requirements.txt
python-resize-image==1.1.20
requests==2.28.2
prettytable==3.7.0
python-dotenv==1.0.0
psycopg==3.1.8
Flask==2.3.1
